created templates for customer and ticket, updated vehicle and controller class
