+++ src/main/java/com/valetparkingtracker/enterprise/dao/CustomerDAOStub.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.dao;
+
+public class CustomerDAOStub {
+}
