+++ src/main/java/com/valetparkingtracker/enterprise/dao/ICustomerDAO.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.dao;
+
+public interface ICustomerDAO {
+}
