+++ src/main/java/com/valetparkingtracker/enterprise/dao/TicketDAOStub.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.dao;
+
+public class TicketDAOStub {
+}
