+++ src/main/java/com/valetparkingtracker/enterprise/dto/Customer.java
@@ -1,0 +1,11 @@
+package com.valetparkingtracker.enterprise.dto;
+
+import lombok.Data;
+
+@Data
+public class Customer {
+    private String customerId;
+    private String firstName;
+    private String lastName;
+    private String ticketId;
+}
