+++ src/main/java/com/valetparkingtracker/enterprise/dto/Vehicle.java
@@ -4,12 +4,10 @@ import lombok.Data;
 
 public @Data
 class Vehicle {
-    private String id;
-    private String firstName;
-    private String lastName;
+    private String vehicleId;
     private String make;
     private String model;
     private String color;
-    private String parkingSpot;
     private String notes;
+    private String ticketId;
 }
