+++ src/main/java/com/valetparkingtracker/enterprise/service/CustomerServiceStub.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.service;
+
+public class CustomerServiceStub {
+}
