+++ src/main/java/com/valetparkingtracker/enterprise/service/ICustomerService.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.service;
+
+public interface ICustomerService {
+}
