+++ src/main/java/com/valetparkingtracker/enterprise/service/TicketServiceStub.java
@@ -1,0 +1,4 @@
+package com.valetparkingtracker.enterprise.service;
+
+public class TicketServiceStub {
+}
