+++ src/main/resources/templates/fragments/topnav.html
@@ -11,7 +11,7 @@
             <a class="nav-link active" aria-current="page" href="/">Home</a>
           </li>
           <li class="nav-item">
-            <a class="nav-link" href="/newVehicle">Add Vehicle</a>
+            <a class="nav-link" href="/checkIn">Check In</a>
           </li>
         </ul>
         <form class="d-flex" role="search" action="#" th:action="@{/vehicle}" method="GET">
