+++ src/main/resources/templates/start.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <div th:replace="~{fragments/topnav :: navbar}"></div>
-    Park a car, Pull a car
+    Put list view of all tickets here
     <p th:text="${vehicle}" />
