+++ src/test/java/com/valetparkingtracker/enterprise/CustomerTests.java
@@ -1,0 +1,11 @@
+package com.valetparkingtracker.enterprise;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class CustomerTests {
+
+    @Test
+    public void testCustomerHere() {}
+}
