+++ src/test/java/com/valetparkingtracker/enterprise/TicketTests.java
@@ -1,0 +1,13 @@
+package com.valetparkingtracker.enterprise;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class TicketTests {
+
+    @Test
+    public void testTicketHere() {
+
+    }
+}
