+++ src/test/java/com/valetparkingtracker/enterprise/ValetParkingTrackerApplicationTests.java
@@ -17,62 +17,7 @@ import static org.mockito.Mockito.verify;
 @SpringBootTest
 class ValetParkingTrackerApplicationTests {
 
-    @Autowired
-    private IVehicleService vehicleService;
-    private Vehicle vehicle = new Vehicle();
-
-    @MockBean
-    private IVehicleDAO vehicleDAO;
-
     @Test
     void contextLoads() {
-    }
-
-    @Test
-    void fetchVehicleByID_returnsVolkswagenForID25() throws Exception {
-        givenVehicleDataIsAvailable();
-        whenVehicle25AddedIsVolkswagen();
-        whenSearchVehicleWithID25();
-        thenReturnOneVolkswagenVehicleForID25();
-
-    }
-
-    private void givenVehicleDataIsAvailable() throws Exception {
-        Mockito.when(vehicleDAO.save(vehicle)).thenReturn(vehicle);
-        vehicleService = new VehicleServiceStub(vehicleDAO);
-    }
-
-    private void whenVehicle25AddedIsVolkswagen() {
-        Vehicle volkswagen = new Vehicle();
-        volkswagen.setId("25");
-        volkswagen.setMake("Volkswagen");
-        Mockito.when(vehicleDAO.fetch(25)).thenReturn(volkswagen);
-    }
-
-    private void whenSearchVehicleWithID25() {
-        vehicle = vehicleService.fetchById(25);
-    }
-
-    private void thenReturnOneVolkswagenVehicleForID25() {
-        String make = vehicle.getMake();
-        assertEquals("Volkswagen", make);
-    }
-
-    @Test
-    void saveVehicle_validateReturnVehicleWithMakeAndModel() throws Exception {
-        givenVehicleDataIsAvailable();
-        whenUserCreatesANewVehicleAndSaves();
-        thenCreateNewVehicleRecordAndReturnIt();
-    }
-
-    private void whenUserCreatesANewVehicleAndSaves() {
-        vehicle.setMake("Volkswagen");
-        vehicle.setModel("Golf");
-    }
-
-    private void thenCreateNewVehicleRecordAndReturnIt() throws Exception {
-        Vehicle createdVehicle = vehicleService.save(vehicle);
-        assertEquals(vehicle, createdVehicle);
-        verify(vehicleDAO, atLeastOnce()).save(vehicle);
     }
 }
